Move math in a dedicated module.
